update paths for quizies
